+++ epoll.c
@@ -205,7 +205,7 @@ FUZZ_FUNC(epoll_wait)
     struct epoll_event *events = malloc(len);
     tfetch_mem(tcp, tcp->u_arg[1], len, events);
 
-    r_set rlist[NUM_RET_EPOLL_WAIT] = {{&ret, sizeof(int), "ret", 0, 0},
+    r_set rlist[NUM_RET_EPOLL_WAIT] = {{&ret, sizeof(long), "ret", 0, 0},
                                        FUZZ_SET_ARRAY(events[0].events, "events",
                                                (int)maxevents, sizeof(struct epoll_event)),
                                        FUZZ_SET_ARRAY(events[0].data.fd, "data",
