+++ stat.c
@@ -52,7 +52,7 @@ FUZZ_FUNC(stat)
     tfetch_mem(tcp, tcp->u_arg[1], len, &fetch_stat);
     kernel_long_t ret = tcp->u_rval;
 
-    r_set rlist[NUM_RET_STAT] = {{&ret, sizeof(int), "ret", 0, 0},
+    r_set rlist[NUM_RET_STAT] = {{&ret, sizeof(long), "ret", 0, 0},
                                  FUZZ_SET(fetch_stat.st_dev, "st_dev"),
                                  FUZZ_SET(fetch_stat.st_ino, "st_ino"),
                                  FUZZ_SET(fetch_stat.st_mode, "st_mode"),
@@ -210,7 +210,7 @@ FUZZ_FUNC(lstat)
     tfetch_mem(tcp, tcp->u_arg[1], len, &fetch_stat);
     kernel_long_t ret = tcp->u_rval;
 
-    r_set rlist[NUM_RET_LSTAT] = {{&ret, sizeof(int), "ret", 0, 0},
+    r_set rlist[NUM_RET_LSTAT] = {{&ret, sizeof(long), "ret", 0, 0},
                                  FUZZ_SET(fetch_stat.st_dev, "st_dev"),
                                  FUZZ_SET(fetch_stat.st_ino, "st_ino"),
                                  FUZZ_SET(fetch_stat.st_mode, "st_mode"),
@@ -250,7 +250,7 @@ FUZZ_FUNC(fstat)
     tfetch_mem(tcp, tcp->u_arg[1], len, &fetch_stat);
     kernel_long_t ret = tcp->u_rval;
 
-    r_set rlist[NUM_RET_FSTAT] = {{&ret, sizeof(int), "ret", 0, 0},
+    r_set rlist[NUM_RET_FSTAT] = {{&ret, sizeof(long), "ret", 0, 0},
                                  FUZZ_SET(fetch_stat.st_dev, "st_dev"),
                                  FUZZ_SET(fetch_stat.st_ino, "st_ino"),
                                  FUZZ_SET(fetch_stat.st_mode, "st_mode"),
